refactor: unify the code
