+++ test/setupFiles/webextension.js
@@ -14,33 +14,8 @@ globalThis.navigator = {
 	userAgent: 'node.js',
 };
 
+const { addRandomDelaysForMethods } = require('./utils');
+
 // Add random delays for async operations,
 // to make tests closer to reality
-const originalStorageLocal = chrome.storage.local;
-chrome.storage.local = new Proxy(originalStorageLocal, {
-	get(target, prop) {
-		const object = target[prop];
-		if (typeof object !== 'function') return object;
-
-		const isNeedDelay = ['get', 'set'].includes(prop);
-
-		return (...args) => {
-			if (!isNeedDelay) {
-				return object(...args);
-			}
-
-			return Promise.resolve().then(async () => {
-				function getRandomInt(max = 1) {
-					return Math.floor(Math.random() * max);
-				}
-
-				const delay = getRandomInt(20);
-
-				console.log('Wait a delay', delay);
-				await new Promise((res) => setTimeout(res, delay));
-
-				return object(...args);
-			});
-		};
-	},
-});
+chrome.storage.local = addRandomDelaysForMethods(chrome.storage.local, ['get', 'set']);
